pyproject.toml
[project]
name = "agentic-ai-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "bs4>=0.0.2",
    "chromadb>=1.2.1",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.120.0",
    "ipykernel>=7.0.1",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4",
    "langchain-core>=1.0.1",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=1.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "langserve>=0.3.3",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.2",
    "sse-starlette>=3.0.2",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
    "wikipedia>=1.4.0",
]
